destroyアクションの制限

+++ app/controllers/items_controller.rb
@@ -35,6 +35,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path if current_user.id != @item.user_id
     if @item.destroy
       redirect_to root_path
     else
